사진 추가 기능 구현 완료

+++ app/src/main/java/com/teampophory/pophory/data/repository/fake/FakePhotoRepository.kt
@@ -51,7 +51,7 @@ class FakePhotoRepository : PhotoRepository {
     override suspend fun addPhoto(
         albumId: Int,
         takenAt: String,
-        studioId: Int,
+        studioId: Long,
         photo: ContentUriRequestBody
     ): Result<Unit> {
         TODO("Not yet implemented")
