+++ app/src/main/java/com/teampophory/pophory/data/repository/photo/DefaultPhotoRepository.kt
@@ -20,7 +20,7 @@ class DefaultPhotoRepository @Inject constructor(
     override suspend fun addPhoto(
         albumId: Int,
         takenAt: String,
-        studioId: Int,
+        studioId: Long,
         photo: ContentUriRequestBody
     ): Result<Unit> {
         return runCatching {
