+++ app/src/main/java/com/teampophory/pophory/network/PhotoNetworkDataSource.kt
@@ -11,7 +11,7 @@ interface PhotoNetworkDataSource {
     suspend fun addPhoto(
         albumId: Int,
         takenAt: String,
-        studioId: Int,
+        studioId: Long,
         photo: ContentUriRequestBody
     )
 }
